requirements.txt
Django==5.2.9
google-genai==1.56.0
langchain-google-genai==4.1.2
langgraph==1.0.5
langchain==1.2.0
langchain-core==1.2.6
python-dotenv==1.2.1
requests==2.32.5
pydantic==2.12.5
gunicorn
whitenoise